pyproject.toml
[project]
name = "discordbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "discord>=2.3.2",
    "discord-py>=2.5.2",
    "ollama>=0.4.8",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.23.5",
    "python-dotenv>=1.1.0",
]

[tool.pytest]
asyncio_mode = "auto"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
markers = [
    "asyncio: mark a test as an asyncio test",
]
